requirements.txt
a2wsgi==1.10.8
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
av==14.0.1
beautifulsoup4==4.12.3
bentoml==1.3.19
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
cattrs==23.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deepmerge==2.0
Deprecated==1.2.15
diffusers==0.32.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-ws==0.7.1
huggingface-hub==0.27.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.31.0
ipywidgets==8.1.5
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kantoku==0.18.1
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
narwhals==1.21.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
open3d==0.19.0
openai==1.59.2
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-aiohttp-client==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
pip-requirements-parser==32.0.1
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.0
questionary==2.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
safetensors==0.5.2
schema==0.7.7
shellingham==1.5.4
simple-di==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli_w==1.1.0
torch==2.5.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
typer==0.15.1
types-aiofiles==24.1.0.20241221
types-beautifulsoup4==4.12.0.20241020
types-html5lib==1.1.11.20241018
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.5.18
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wrapt==1.17.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
